GUI Toolbar and Data Object
The GUI has a toolbar now with a start battle and exit button. We also
have new icons now.

In terms of data objects thus far, we need to come up with a database of
some sort to host the data for the project. We also need to complete the
methods for combatants as well as create a class for "moves." Finally,
we will need to develop algorithms for a min-max tree.