EasyAI Abandon Notice
I managed to get the basic classes working with easyAI, BUT the AI
algorithm is garbage and needs to be rewritten. The AI kept going for
the first move in the list and wouldn't even try new moves. Therefore,
I'm backing up the current project and then writing my own classes.
Documentation will be updated as appropriate.