Successful Test
There are some errors in execution listed in the To-Do List.txt file,
but, for the most part, the AI tried out various moves, as intended, and
eventually found the ideal move. In the end I won, but only because I
cheated and hit it with the strongest move straight away.