Moved Images Begin GUI Transition
I forgot to move the images to the appropriate directory, so that's been
fixed.

There's still a ways to go for GUI development, but I now know how to
mimic a print statement in the GUI. I won't be converting all print
statements just yet, but I can start creating widgets with moves, guides
and stats. We'll be keeping computer stats hidden with the GUI version,
but they can still be opened in a text editor.